+++ CHANGELOG.md
@@ -15,6 +15,15 @@ our [guidelines for writing a good changelog entry](https://github.com/biomejs/b
 
 ### CLI
 
+#### Enhancements
+
+- Biome now executes commands (lint, format, check and ci) on the working directory by default. [#2266](https://github.com/biomejs/biome/issues/2266) Contributed by @unvalley
+
+  ```diff
+  - biome check .
+  + biome check    # You can run the command without the path
+  ```
+  
 ### Configuration
 
 ### Editors